Add files via upload
created using HTML,CSS,PYTHON,PYTHON FLASK,MYSQL. This project implements the blockchain for a simple small voting system. explaining flow of process the login credentials are checked with database and if login passes it will redirect to voter page else it will display as invalid credential page. the voter page redirect tocoting page where voter can vote for parties and only one vote per voter is allowed to avoid double voting ,this is checked using a table in database by updating status of vote. if the voter try to vote again it will direct to already voted page from there we can go to miner page . the "mine" route may represent the action taken by a miner node to add a new block to the blockchain. This block would contain the recent transactions (votes) that have been verified and are ready to be added to the blockchain.the "chain" route may be used by voters or other nodes to view the entire history of votes and ensure the transparency and integrity of the voting process.